refactor: выделил классы для настройки и обработки данных
Созданы отдельные классы для работы с настройками, студентами, заданиями и подачами, что улучшает читаемость кода и его структуру. Это также упрощает управление настройками и данными в проекте.
